requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
email-validator==2.1.0
Pillow==10.1.0 